Convert dataframes to float before assignment

+++ python/core/original_calculations.py
@@ -99,8 +99,9 @@ def get_PipeSize_Suggested(F1):
         # print(f"📊 PIPSZ data shape: {pipsz_df.shape}")
         
         # Convert columns to numeric
-        pipsz_df.iloc[:, 0] = pipsz_df.iloc[:, 0].apply(universal_float_convert).astype(float)  # Flow capacity
-        pipsz_df.iloc[:, 1] = pipsz_df.iloc[:, 1].apply(universal_float_convert).astype(float)  # Pipe size
+        pipsz_df = pipsz_df.astype(float)  # Convert DataFrame to float
+        pipsz_df.iloc[:, 0] = pipsz_df.iloc[:, 0].apply(universal_float_convert)  # Flow capacity
+        pipsz_df.iloc[:, 1] = pipsz_df.iloc[:, 1].apply(universal_float_convert)  # Pipe size
         
         # Remove invalid rows
         valid_rows = pipsz_df.dropna()
@@ -155,8 +156,9 @@ def get_PipeLength(F1, T1, T2):
             
         # Convert to numeric and find appropriate length
         room_df = room_df.copy()
-        room_df.iloc[:, 0] = room_df.iloc[:, 0].apply(universal_float_convert).astype(float)  # Power capacity
-        room_df.iloc[:, 1] = room_df.iloc[:, 1].apply(universal_float_convert).astype(float)  # Length
+        room_df = room_df.astype(float)  # Convert DataFrame to float
+        room_df.iloc[:, 0] = room_df.iloc[:, 0].apply(universal_float_convert)  # Power capacity
+        room_df.iloc[:, 1] = room_df.iloc[:, 1].apply(universal_float_convert)  # Length
         
         # Find ceiling match
         adequate_rows = room_df[room_df.iloc[:, 0] >= power_mw]
@@ -386,8 +388,9 @@ def get_system_sizing(system_data):
         if room_df is not None:
             room_df = room_df.copy()
             # Convert columns to numeric
-            room_df.iloc[:, 0] = room_df.iloc[:, 0].apply(universal_float_convert).astype(float)  # Power column
-            room_df.iloc[:, 1] = room_df.iloc[:, 1].apply(universal_float_convert).astype(float)  # Room size column
+            room_df = room_df.astype(float)  # Convert DataFrame to float
+            room_df.iloc[:, 0] = room_df.iloc[:, 0].apply(universal_float_convert)  # Power column
+            room_df.iloc[:, 1] = room_df.iloc[:, 1].apply(universal_float_convert)  # Room size column
             
             for idx, row in room_df.iterrows():
                 power_val = row.iloc[0]
